requirements.txt
asgiref==3.9.2
black==25.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
Django==5.2.6
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
flake8==7.3.0
idna==3.10
mariadb==1.1.13
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.1.1
pytokens==0.1.10
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.3
tweepy==4.16.0
tzdata==2025.2
urllib3==2.5.0
